Configured clang-format to group and sort includes
As clang 6 is in broad usage already, it's time to use its features.
This change enables clang-format to automatically apply the rules from
our coding guidelines (this is a quote from docs/General.md):

---
The style used to specify an include shall reflect the locality of the
included header file.

1. All includes which are part of the module the including header file
is part of shall use quotes ""
2. All includes which are not part of the module the including header
(external includes) file is part of shall use angle brackets <>
3. Sort includes alphabetically
4. Put external includes after internal ones
---

I also took the liberty to separate platfrom and estd headers as a
separate group - placing them between all the library headers in <> and
standard library headers.